requirements.txt
authlib>=1.0
Flask==2.0.1
Werkzeug==2.0.3
spotipy==2.19.0
requests==2.26.0
python-dotenv==0.19.0